feat: improve theme example
This change starts to move this package from being a test theme to more
of a proper example theme that is sufficiently different from the
default Developer Hub theme.  This change renames the theme plugin name
and updates the theme colors, as well as a couple component overrides to
make it much more obvious that the theme is loaded.  This commit also
simplifies the readme and steps to view the example.  Finally this
change also adds a plugin development setup and starts some simple pages
to experiment with theme changes.

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>